// src/lib/api-client/models/config-dnsvalidator-config-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Domain Flow API
 * A comprehensive API for domain generation, validation, and management operations.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ConfigDNSValidatorConfigJSON
 */
export interface ConfigDNSValidatorConfigJSON {
    /**
     * 
     * @type {number}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'concurrentQueriesPerDomain'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'maxConcurrentGoroutines'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'maxDomainsPerRequest'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'queryDelayMaxMs'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'queryDelayMinMs'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'queryTimeoutSeconds'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'rateLimitBurst'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'rateLimitDps'?: number;
    /**
     * 
     * @type {string}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'resolverStrategy'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'resolvers'?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'resolversPreferredOrder'?: Array<string>;
    /**
     * 
     * @type {{ [key: string]: number; }}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'resolversWeighted'?: { [key: string]: number; };
    /**
     * 
     * @type {boolean}
     * @memberof ConfigDNSValidatorConfigJSON
     */
    'useSystemResolvers'?: boolean;
}