// src/lib/api-client/models/config-httpvalidator-config-json.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Domain Flow API
 * A comprehensive API for domain generation, validation, and management operations.
 *
 * The version of the OpenAPI document: 1.0
 * Contact: support@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ConfigHTTPValidatorConfigJSON
 */
export interface ConfigHTTPValidatorConfigJSON {
    /**
     * 
     * @type {boolean}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'allowInsecureTLS'?: boolean;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'defaultHeaders'?: { [key: string]: string; };
    /**
     * 
     * @type {string}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'defaultUserAgent'?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'followRedirects'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'maxBodyReadBytes'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'maxConcurrentGoroutines'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'maxDomainsPerRequest'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'maxRedirects'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'rateLimitBurst'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'rateLimitDps'?: number;
    /**
     * 
     * @type {number}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'requestTimeoutSeconds'?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof ConfigHTTPValidatorConfigJSON
     */
    'userAgents'?: Array<string>;
}